feat: searchMovieList를 movieList로 통합, 포스터 이미지가 없는 예외 처리
