--- 코튼_김연진/6주차/ts-movie/src/components/searchMovieList/searchMovieList.style.ts
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-const Container = styled.div`
-    display: flex;
-    width: 95%;
-`;
-
-const Text = styled.div`
-    margin-top: 30px;
-    color: white;
-    font-size: 30px;
-`;
-
-export { Container, Text };
