+++ 코튼_김연진/6주차/미션2/src/components/movieCard/movieCard.style.js
@@ -10,6 +10,7 @@ const Container = styled(Link)`
         width: 140px;
         height: 200px;
         border-radius: 10px;
+        object-fit: cover;
     }
 `;
