+++ 코튼_김연진/6주차/미션2/src/components/movieList/movieList.jsx
@@ -16,7 +16,7 @@ const MovieList = ({ url, mq }) => {
         return <Error />;
     }
 
-    if (movies?.data?.results?.length === 0) {
+    if (mq && movies?.data?.results?.length === 0) {
         return (
             <S.Container>
                 <S.Text>
