--- 코튼_김연진/6주차/미션2/src/components/searchMovieList/SearchMovieList.style.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-const Container = styled.div`
-    flex: 1;
-    width: 100%;
-`;
-const Text = styled.div`
-    margin-top: 30px;
-    color: white;
-    font-size: 30px;
-`;
-
-export { Container, Text };
