+++ 코튼_김연진/6주차/미션2/src/pages/search/search.jsx
@@ -1,7 +1,7 @@
 import * as S from './search.style';
 import { useState } from 'react';
 import { useSearchParams, useNavigate } from 'react-router-dom';
-import { SearchMovieList } from '../../components/searchMovieList/SearchMovieList';
+import MovieList from '../../components/movieList/movieList';
 
 const Search = () => {
     const [searchValue, setSearchValue] = useState('');
@@ -14,7 +14,7 @@ const Search = () => {
         mq: '',
     });
 
-    const mq = searchParams.get('mq');
+    const mq = searchParams.get('mq') || '';
 
     const handleSearchMovie = () => {
         if (mq === searchValue) return;
@@ -38,7 +38,10 @@ const Search = () => {
                 ></S.SearchBox>
                 <S.Button onClick={handleSearchMovie}>검색</S.Button>
             </S.Search>
-            <SearchMovieList></SearchMovieList>
+            <MovieList
+                url={`/search/movie?query=${mq}&include_adult=false&language=ko-KR&page=1`}
+                mq={mq}
+            ></MovieList>
         </S.Container>
     );
 };
